Componentize HasIbcChainTypes
